make FrostedGlassBackground a public reusable view
